Use one page buffer per column rather than one buffer per page
Replace the columnWriter pages slice of buffers with a single pageBuffer
that contains data for all pages in the row-group.

This addresses the fact that when using file buffers, large numbers of
files are held open, which can cause errors. It also places more data in
contiguous chunks, which should make writing faster.

When using in-memory buffers, this should cause fewer (but larger)
buffer reallocations. I expect this to be good on-net, but if the old
characteristics are desirable that can be addressed by changing the
in-memory buffer implementation.