Extract zero values when encoding nil `any` values
This commit adds nil checks to both interface and underlying value types whilst
encoding such that `any` values provided as non-optional columns are encoded as
zero values.

This is arguably a better behaviour than panicking. However, the provided values
in these cases could be considered invalid as input to the provided schema as
the columns are not marked as optional. If we were to refactor these encoding
functions to provide error messages I would suggest we consider returning errors
here instead of continuing with zero values.

Signed-off-by: Ashley Jeffs <ash@jeffail.uk>