Add alternative in-memory BufferPool implementations
Add chunkMemoryBufferPool and allocMemoryBufferPool as alternative
buffer pool implementations with better performance than the regular
memoryBufferPool as the column buffer pool.

Both buffer implementations maintain a slice of byte slices rather than
a single contiguous byte slice to avoid reallocs and copies when the
buffer fills. Both buffers use a sync.Pool of byte-slices.

chunkMemoryBuffer uses fixed-size slices to reduce the number
allocations, provide large contiguous chunks of memory for reading, and
ensure pooled buffers are exactly the right size. By comparison
allocMemoryBuffer pulls a new buffer from the pool for every call to
Write, which may need to be reallocated if it doesn't contain enough
space.

Initially, chunkMemoryBufferPool was intended to be the default
in-memory buffer implementation for the column buffer pool, but in
practice chunkMemoryBuffer performs worse than allocMemoryBuffer for
unclear reasons. allocMemoryBuffer was designed to work nearly the same
(and have nearly the same performance) as when a separate buffer was
generated for each page.