add tagSource interface for providing parquet tags
Part of #44

This can be reviewed and merged separately

This commit adds `tagSource` interface that defines api for providing parquet
tags from `*reflect.StructField`.

The api is private to avoid user provided implementation. We reserve
interpretation and processing of the tags as an implementation detail.

The api explicitly passes tagSource instead of variadic ...tagSource to ensure
that processing should not care about where the tags come from.

For now defaultTagSource is the only implementation that reads `parquet`
`parquet-key` and `parquet-value` tags