Lazy-load column index
The parquet.OpenFile allows opening a parquet file by either parsing the entire
page index, or not parsing it at all. The cost of the former can be significant when
opening large files.

This commit adds lazy parsing for column and offset indices when they are read from file-backed chunks.
The advantage is that the page index can be skipped when opening a file, and column and offset indices can
still be accessed on-deman when requested through individual chunks.