Fix NumValues for missingColumnChunk
A missingColumnChunk can be created during schema conversion when
the column does not exist in the target schema. The number of values
for the chunk is considered to be equal to the number of rows in the
row group, but the value returned by NumValues is currently 0.

Since the page of the chunk returns the number of rows in the group as
number of values, it would be consistent for the chunk to do the same.