Fix column index reuse (#92)
Previously the base column index would return a pointer to the nullpages
and nullcounts slices when generating a column index. However if the
index was reset and reused these pointers would change the original
value of the generated index.

This fixes that by copying the byte slices when generating the index
instead of just using the slice pointers