Return errors for missing page indices
As a result of https://github.com/parquet-go/parquet-go/pull/84, we can now return
a nil Column or Offset index when the SkipPageIndex is set. Furthermore, if the page
index is not skip when opening the file, the return value will be a zero-value struct
instead of nil.

This commit reconciles this inconsistency by returning `ErrMissingColumnIndex` and `ErrMissingOffsetIndex`
respectively when either index is missing.

Follow up to https://github.com/parquet-go/parquet-go/pull/85#discussion_r1400269216.