Pool column index buffers when performing comparisons
This commit adds a pool to allocate column index slices during comparisons.
Production profiles have shown these slices to account for 10% of our
allocations in terms of space.

The pool is relatively simple. It returns nothing but a fixed-size slice
regardless of the number of columns being processed to avoid pool
fragmentation. The capacity of these slices have been bumped up to 128 from 64
given that it allows for a higher number of columns (beneficial to us
specifically) with only a 512B increase per allocation which should be
amortized by the pool now anyway.