Estimate dictionary decode size
As originally described in https://github.com/parquet-go/parquet-go/pull/96, column dictionaries
are not properly sized befored decoding which leads to many small allocations in a loop.

This PR extracts the part of #96 which does buffer size estimation for dictionaries
until we figure out a safe way to pool those buffers.