Remove allocMemoryBufferPool implementation
Further testing showed that the chunkMemoryBufferPool only performed
badly in very specific scenarios (i.e. on one machine under one
operating system). Under most circumstances, chunkMemoryBufferPool and
allocMemoryBufferPool both perform similarly to the old code.

Because there are likely benefits to using the chunkMemoryBufferPool
over the allocMemoryBufferPool, the former is kept and made the new
default for column the column buffer pool and the latter is removed.