search: fix bug in binary search on overlapping page bounds
If we have overlapping page bounds because of truncated pages the binary
search algorithm doesnt find the first page that could contain the value
anymore.
This happens because it recurses to the right if the value is not
smaller then the min value, but in degenerate cases it might be never
smaller then the min value because all pages have same bounds.
We fix that by iterating pages backwards until we find the smallest
page, accounting for overlap.

Credit goes to @alanprot for finding the issue and writing the test.

Signed-off-by: Michael Hoffmann <mhoffmann@cloudflare.com>
Signed-off-by: Michael Hoffmann <mhoffm@posteo.de>