Add a Close method on the column writer
When writing one column at a time, it is useful to close columns as
writes are finished in order to release underlying resources.

For that purpose, this PR implements io.Closer on the ColumnWriter type
which will flush the data buffer and set it to nil so it can be garbage collected.