Pool buffers for optional column writer.
The optionalColumnBuffer uses direct allocation for tracking rows and definition levels.
This can become a problem when writing wide, sparse tables with many nulls as memory can
grow very quickly.

This commit adds pooling for the underlying buffers by refactoring the existing buffers
implementation to work with type parameters. Byte buffers are defined as a generic type alias
which reduces the need to do large refactoring.