Fix memory leak in CopyPages
The CopyPages function does not release pages after it is done with
copying values to the page writer. Although this can release the underlying bytes
for reuse, my understanding is that the writer should clone values it intends to
retain because of this exact reason.

The issue can be reproduced by running the added test without the added parquet.Release call:
PARQUETGODEBUG=tracebuf=1 go test . -v -count 10 -run TestCopyFilePages