writer: support concurrent row group construction
There are times when creating very large parquet files when the
bottleneck is creating the files themselves. Currently a lot of the work
in the existing APIs can be parallelized (for example record shredding
can be done externally, etc), but the final step of actually serializing
the values and compressing the rows must be done serially.

This change introduces new APIs that allows writing different row groups
concurrently, as long as they are "committed" back to the original
writer serially. This allows for offloading expensive work (like
compression, etc) that has been shown to be a bottleneck to multiple
threads, as the last step is mostly just moving/copying around some
buffers.

Currently in Redpanda Connect, we parallelize the construction of
parquet files in some places, but this requires we hold the entire
matrix of parquet.Value in memory, which is very expensive. This patch
will allow us to reduce the required buffering, and instead we keep
around compressed pages for each row group in memory instead.