support mapping json messages to parquet columns (#363)
Continuing on #358, this PR adds support for writing json payloads
directly to parquet without requiring an intermediary conversion to a
json value.

Technically, the conversion still happens, because the trade off is
either loading some form of in-memory representation of the json
document, or parsing it multiple times as we walk through the nested
values. In practice, the former is a better approach because doing
multiple pass over nested json fields exhibits O(N^2) behavior.

At the API layer, parquet-go now interprets `json.Number` and
`json.RawMessage`, mapping the content of those values to parquet
columns.

The in-memory representation is optimized to minimize heap allocations
(compared to using `json.Unmarshal` into `any` for example). ~I chose to
use https://github.com/json-iterator/go for the the parser because it's
a well tested and wildly used package. It has shortcomings tho, for
example, reading string values from the underlying json causes heap
allocations for each value. At this stage, I'm interested in getting the
feature integrated, we can always improve efficiency later.~ I removed
the dependency and addressed the low hanging fruits on the performance
side.

As part of this change, I also had to enhance the support for date and
time logical types to be able to use it in tests where json values were
converted to parquet columns of those types.

---------

Signed-off-by: Achille Roussel <achille.roussel@gmail.com>