fix buffer reuse bug (#374)
Changes from #372 introduced a misused of the internal buffer pool,
causing buffer reuse issues where the content of a previous use was
still visible in a subsequent write.

I addressed the issue by removing the use of buffer pools in the column
writer path in favor of local buffers that get reused; this solution
works because writes to column buffers are not concurrent.

---------

Signed-off-by: Achille Roussel <achille.roussel@gmail.com>