Fix size statistics (#385)
We had a TODO in `format/parquet.go` about the presence of size
statistics causing errors in the tests opening files from `testdata`. It
turned out that apparently we had a non-compliant file that had invalid
content causing the parser to fail. I moved the problematic file into
`deprecated/testdata` and documented the issue, then implemented the
missing size statistics.

---------

Signed-off-by: Achille Roussel <achille.roussel@gmail.com>