Fix issue 356 (#388)
Fixes #356 

This bug turned out to be a more profound problem caused by non-equal
but compatible schemas configured on `GenericWriter[T]`.

I refactor the internals to remove the way we historically relied on
`parquet.Schema.Deconstruct`, and replaced it with the newly added
reflection-based writes from `column_buffer_reflect.go`. This unifies
the writing behavior across the board, it should prevent a wider range
of bugs like the one that was originally reported.

---------

Signed-off-by: Achille Roussel <achille.roussel@gmail.com>